+++ lizard_auth_server/__init__.py
@@ -1,1 +1,2 @@
 # package
+default_app_config = 'lizard_auth_server.apps.MyAppConfig'