+++ CHANGES.rst
@@ -5,6 +5,12 @@ Changelog of lizard-auth-server
 1.8 (unreleased)
 ----------------
 
+- Added new models for SSO refactoring.
+
+- Put user creation signals handlers into seperate module.
+
+- Some py3 changes.
+
 - Renamed 'return_unauthenticated'.
 
 - A user arriving at the SSO server after being redirected there can
@@ -23,7 +29,6 @@ Changelog of lizard-auth-server
   workflow that is sometimes helpful.
 
 - Python 3 setup and test fixes.
-
 
 1.7 (2016-06-14)
 ----------------